mybot/requirements.txt
Django>=5.2.0,<5.3.0
celery>=5.5.0,<5.6.0
redis>=6.2.0,<6.3.0
django-celery-beat>=2.8.0,<2.9.0
botbuilder-core>=4.17.0,<4.18.0
botbuilder-schema>=4.17.0,<4.18.0
botbuilder-integration-aiohttp>=4.17.0,<4.18.0
botbuilder-azure>=4.17.0,<4.18.0
aiohttp>=3.12.0,<3.13.0
gunicorn>=23.0.0,<24.0.0
requests>=2.31.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
asgiref>=3.8.1,<3.9.0
pytz>=2024.1 
openai>=1.0.0